+++ cohttp-eio/src/server.ml
@@ -1,5 +1,6 @@
-open Eio.Std
-module Write = Eio.Buf_write
+module Buf_read = Eio.Buf_read
+module Buf_write = Eio.Buf_write
+module Switch = Eio.Switch
 
 type middleware = handler -> handler
 and handler = request -> response
@@ -15,14 +16,13 @@ let domain_count =
 
 let read_fixed request reader =
   match Http.Request.meth request with
-  | `POST | `PUT | `PATCH -> Body.read_fixed reader request.headers
-  | _ ->
-      let err =
-        Printf.sprintf
-          "Request with HTTP method '%s' doesn't support request body"
-          (Http.Method.to_string request.meth)
-      in
-      raise @@ Invalid_argument err
+  | `POST | `PUT | `PATCH ->
+      let ( let* ) o f = Option.bind o f in
+      let ( let+ ) o f = Option.map f o in
+      let* v = Http.Header.get request.headers "Content-Length" in
+      let+ content_length = int_of_string_opt v in
+      Buf_read.take content_length reader
+  | _ -> None
 
 let read_chunked request reader f =
   Body.read_chunked reader (Http.Request.headers request) f
@@ -65,26 +65,41 @@ let internal_server_error_response =
 let bad_request_response =
   (Http.Response.make ~status:`Bad_request (), Body.Empty)
 
-let write_response (writer : Write.t)
-    ((response, body) : Http.Response.t * Body.t) =
+let write_response writer ((response, body) : Http.Response.t * Body.t) =
   let version = Http.Version.to_string response.version in
   let status = Http.Status.to_string response.status in
-  Write.string writer version;
-  Write.string writer " ";
-  Write.string writer status;
-  Write.string writer "\r\n";
-  Body.write_headers writer response.headers;
-  Write.string writer "\r\n";
-  match body with
-  | Fixed s -> Write.string writer s
-  | Chunked chunk_writer -> Body.write_chunked writer chunk_writer
-  | Custom f -> f writer
-  | Empty -> ()
+  Buf_write.string writer version;
+  Buf_write.char writer ' ';
+  Buf_write.string writer status;
+  Buf_write.string writer "\r\n";
+  Rwer.write_headers writer response.headers;
+  Buf_write.string writer "\r\n";
+  Body.write_body writer body
+
+(* request parsers *)
+
+let meth =
+  let open Eio.Buf_read.Syntax in
+  let+ meth = Rwer.(token <* space) in
+  Http.Method.of_string meth
+
+let resource =
+  let open Eio.Buf_read.Syntax in
+  Rwer.(take_while1 (fun c -> c != ' ') <* space)
+
+let[@warning "-3"] http_request t =
+  let open Eio.Buf_read.Syntax in
+  let meth = meth t in
+  let resource = resource t in
+  let version = Rwer.(version <* crlf) t in
+  let headers = Rwer.http_headers t in
+  let encoding = Http.Header.get_transfer_encoding headers in
+  { Http.Request.meth; resource; version; headers; scheme = None; encoding }
 
 (* main *)
 
 let rec handle_request client_addr reader writer flow handler =
-  match Reader.http_request reader with
+  match http_request reader with
   | request ->
       let response, body = handler (request, reader, client_addr) in
       write_response writer (response, body);
@@ -99,10 +114,8 @@ let rec handle_request client_addr reader writer flow handler =
       raise ex
 
 let connection_handler (handler : handler) flow client_addr =
-  let reader =
-    Eio.Buf_read.of_flow ~initial_size:0x1000 ~max_size:max_int flow
-  in
-  Write.with_flow flow (fun writer ->
+  let reader = Buf_read.of_flow ~initial_size:0x1000 ~max_size:max_int flow in
+  Buf_write.with_flow flow (fun writer ->
       handle_request client_addr reader writer flow handler)
 
 let run_domain ssock handler =